+++ app/src/main/java/org/mozilla/fenix/settings/quicksettings/QuickSettingsComponent.kt
@@ -86,7 +86,7 @@ class QuickSettingsComponent(
                 PhoneFeature.MICROPHONE -> microphone = microphone.toggle()
                 PhoneFeature.NOTIFICATION -> notification = notification.toggle()
             }
-            context.components.storage.addSitePermissionException(origin, location, camera, microphone, notification)
+            context.components.storage.addSitePermissionException(origin, location, notification, microphone, camera)
         } else {
             val updatedSitePermissions = sitePermissions.toggle(featurePhone)
             context.components.storage.updateSitePermissions(updatedSitePermissions)